+++ src/parse/mod.rs
@@ -9,3 +9,4 @@ pub mod cytobands;
 pub mod genetic_models;
 pub mod info;
 pub mod strs;
+pub mod meis;