+++ CHANGELOG.md
@@ -27,5 +27,6 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 - Parse a number of STR-related fields (#21)
 - Parse SCOUT_CUSTOM INFO field (#22)
 - Parse strs source fields (`SourceDisplay`, `Source`, `SourceId`) (#23)
+- Parse mei info: `name` and `polarity` (#24)
 ### Fixed
 - Normalize chromosome names by stripping the chr prefix (#4)